int->long

+++ simple_benchmark/simple_benchmark.c
@@ -3,9 +3,9 @@
 
 int main(int argc, char** argv) {
     if (argc != 2) return 1;
-    int n = atoi(argv[1]);
+    long n = atol(argv[1]);
 
-    for (int k = 0; k < n; k++) {
+    for (long k = 0; k < n; k++) {
         asm("nop");
     }
